Removed #define TXF_COMMIT_METHOD
The default is now TXF_BACKOUT_METHOD.  The possibility to switch
it around using -D parameters and using TXF_NO_BACKOUT_METHOD
or TXF_NO_COMMIT_METHOD remains in place, but right now a problem
was found in trying to use this repo for using the COMMIT method.
Hence, until further notice, only use this repo in the default manner
for the BACKPUT method only.