Add TXF_BACKOUT_METHOD and TXF_COMMIT_METHOD
A #define for both exist by by default, but either can be disabled with
a TXF_NO_BACKOUT_METHOD or TXF_NO_COMMIT_EMTHOD, so that
either the original COMMIT implementation, or the alternative
BACKOUT implementation can be chosen.  When both are enabled,
the original COMMIT method prevails, and the BACKOUT only performs
some verifications.